added powers and origin templates to data ao that they can be accessed normally and are part of the base origins. also added origins to mod dependencies, so they are grouped together now.

+++ Toads Simple Mods/Fabric-TSOMod-FishBowl-1.21.1/src/main/java/net/spottedtoad/toads_simple_origins/item/custom/FilledFishBowlArmorItem.java
@@ -58,7 +58,7 @@ public class FilledFishBowlArmorItem extends ArmorItem {
                         int currentDamage = itemStack.getDamage();
                         NbtComponent nbtComponent = itemStack.getOrDefault(DataComponentTypes.CUSTOM_DATA, NbtComponent.DEFAULT);
                         NbtCompound itemNbt = nbtComponent.copyNbt();
-                        int currentOxygen = itemNbt.contains("oxygenLevel") ? itemNbt.getShort("oxygenLevel") : maxOxygen;
+                        int currentOxygen = itemNbt.contains("oxygenLevel") ? itemNbt.getInt("oxygenLevel") : maxOxygen;
                         //Apply durability and oxygen data
                         BlockEntity placedEntity = world.getBlockEntity(placePos);
                         if (placedEntity instanceof FilledFishBowlBlockEntity filledEntity) {
